Set up skeleton for "my tagged traits" on profile page
Set up a new tab on the profile template for displaying traits that a user has
tagged. Set up the UserTaggedTraitTable class and a test for it. I was going
to set up the whole thing, but getting a list of traits that a user has tagged
requires going through the UserData model, which is difficult now because it's
still set up wrong (should be a OneToOneField, but it's not). So I'll leave that
to do when I fix up the UserData model.