Prevent users from repeating traits in two trait fields of tagging forms
For the ManyTaggedTraits form classes, add logic to the form's clean() methods
to make the form invalid if a user specifies the same trait in both the 'traits'
and 'recommended traits' fields. As it was, I think that if a trait was repeated
it would have been tagged with recommended=False, and then it would potentially
prevent tagging of other recommended traits (any that came after it in the list
of recommended traits).