Add delete buttons to tagged trait study list
On the page with a list of tagged traits for each study, display a different
version of the table, with delete buttons, if the logged in user is a tagger
for the study in question. Add another table class for this, as well as logic
in the view for selecting the correct table class (override get_table_class()
from the generic table view mixin from django-tables2). Add tests of the
conditional table class choice.