Cleaning up tests after making profile creation automatic
Many tests and/or TestCase superclasses created a user and then created a profile
for that user. These tests failed now that Profile.user is a OneToOneField because
it created a **second** profile for the user, after one was automatically created
for the user. So remove all of the code that created duplicate user profiles, and
make sure the tests pass.