Really fix group/permission data migrations this time
Rewrite data migrations for groups and permissions. Use the Group and Permissions
classes instead of any data fixtures. Add dependencies for the migrations. Add
a call to emit_post_migrate_signal() so that content types and permissions are
created before any of the groups and permissions are needed in these migrations.
Also add some dependencies to other migrations, now that I've thought more about what
the dependencies mean. Also add tests that check the groups and permissions that
I expect to exist.