Add trait count to StudyTable
Add trait_count column to the StudyTable, to display the count of non-deprecated traits
for a given study. Add tests for this rendered column to the trait_browser.test_tables
test file. Also add some other tests that don't do much, but were practice for figuring
out how to write the other test.