Fix TaggedTrait creation form behavior for previously tagged traits
Previous behavior was to throw an error (resulting in an HTTP 500 server error
page on the production site for users) if any of the TaggedTrait creation forms
received data for creating a duplicate TaggedTrait. To prevent this, add clean
methods to the single TaggedTrait forms and add functionality to the clean methods
of the ManyTaggedTrait forms. Also add a clean method to the TaggedTraitAdmin form.
Add tests of the new form errors, and tests of the views. Since this necessitated
adding the tag as an attribute of the ByTag forms, also remove the weird way of
displaying the tag name in the form template for the ByTag forms, and instead add
the tag name to the subtitle form attribute in the forms' __init__ methods.