Fix error reporting for creating tags via admin interface
Add a clean method to the TagAdminForm to handle the case where a title is input
that results in a non-unique lower_title. Previously, this situation threw a ValidationError,
which results in users of the production site seeing a 500 server error page. Now they
get an error message on the form instead.