Remove the recommended field from tags.TaggedTrait model
Remove the recommended field from the tags.TaggedTrait model. Add a new migration
to effect this change. Remove references to the recommended field in all of the
other parts of the tag app. The most complicated removal procedure was in the
forms and views, along with their corresponding test files.