Refine study detail and dataset detail views
Redesign study detail page to match styling of dataset detail page. Add a table
of variables and a table of datasets to the study detail page. This is currently
working strangely, since pagination changes the page for both tables at once,
so this is likely to change in the next commit. Also tweak the SourceDatasetTable
classes.