Improve test method names for autocomplete views
For tests of the source trait autocomplete views, improve the names of
the test methods, making them more accurate and more uniform. This will
make it easier when I add tests for the new autocomplete behavior next.
Also move the test about case-insensitivity for name autocompletion next
to the other test about the name.