Fix phv autocomplete with leading zeros
Assume that the query string is always started from the beginning, and
use that assumption to fix the auto-completion of source traits by phv
when leading zeros are present. This involves an if/else block that does
one kind of source trait filter for query strings that start with a 0
and a different kind of filter for query strings that start with a
non-zero digit.