Don't return deprecated harm. traits in search
Add exclusion of deprecated harmonized traits in the search function.
But don't add a test for this, since the search functions are about to
be replaced anyway. Change HarmonizedTraitSetVersionFactory to make
i_is_deprecated field False by default (was previously a Faker boolean,
which made tests of the harmonized trait search function randomly fail
50% of the time with this new change).