Fill in tests for SourceTraitSearchView
Fill in test skeleton methods for the SourceTraitSearchView. Add a
tearDown method similar to the one in test_searches.py that deletes
all existing SearchEntry objects. Add a new test for the view when
an invalid (blank) form has been submitted. Note that all tests may
not yet be passing.