Split get processing into form_valid and form_invalid methods
Move the conditional statements for form validity into form_valid
and form_invalid methods. Render the response in these methods
instead of in the get method. Set the default table_data to nothing
and only set it to a valid set of traits if the form is valid in the
form_valid method.