Use messages to report number of search results
Use the MessagesMixin from django-braces to pass an info message
with the number of results to the template. Set the view's message
attribute when the form is valid. Include the _messages.html
template in the source trait search template so that the messages
are properly displayed. Add tests that check that the proper
message (including pluralization) is displayed upon a valid search,
and that no messages are displayed otherwise.