Add a study field to the SourceTraitSearchForm
Add a study field to the form, using the autocompletion widget.
Update existing form test names and docstrings to indicate that they
do not include studies in the form. Add new tests that explicitly
check form behavior when provided with existing or non-existent
study pks. Include a footer in the search template so that the
django autocomplete light widget works properly.