Allow searching by exact trait name
Update the search function to accept a trait name argument and
search for traits whose names exactly match that (case insensitive).
Add a field to the search form to capture an optional trait name.
Add tests for the search fucntion, the form, and the view to
verify that only traits whose names exactly match the specified
name are found.