Require either name or description when searching
Update search function to not require the description. Update tests
to reflect this behavior by renaming some with more appropriate
names, removing searches by description in the tests checking trait
name, and adding new tests. Make the q field in the form not
required. Add a clean method to the form to check that either the
trait name or the description are not blank.