Allow searches to be by trait name substring
Add a keyword argument to source_trait_search that specifies if the
specified name string should exactly match the trait name or if it
should just be a substring in the name. Add a checkbox to the form
that allows the user to select this behavior. Add tests to the
test_searches and test_forms files.