Change search form and view based on Leslie's comments
Update names of fields in the form to include "Variable". Add better
help text for the studies field and change the name to indicate more
clearly that either one or multiple studies may be selected. Update
comment in the _search_help.html template to reflect the intended
purpose of the template. Wrap text differently trait_browser app.