Add a reset button to the search form
Add a second submit button labeled as "reset" that will clear
existing search info. Update the SourceTraitSearchView to redirect
to the search page, which clears get parameters. Add tests. Note
that including the css_class 'btn-disable' breaks the intended
behavior because "reset" is no longer passed as a get parameter and
therefore the view acts like it is being searched normally.
(Unfortunately I am not sure why this is happning and so I couldn't
find a way to test this in the unit tests.)