Add a harmonized trait search page
Add a new class-based view for the harmonized trait search,
including a template. Add a form for searching harmonized traits.
Update the urls to point to a the a new page (watsonsearch) that
can later be remapped to the intended search page. Add tests for
all functionality, including the import_db step.