Fix source trait autocomplete for phv in trait name
Rewrite get_queryset() in the two SourceTrait autocomplete by name or
phv views to account for trait names that might start with 'phv'. Now
the autocomplete view will return traits with the right phv accession
AND with trait names starting with 'phv#'. Add tests of the new behavior
and tweak tests that didn't work quite right anymore with the changes.
Also found in issue in
TaggableStudyFilteredSourceTraitNameOrPHVAutocomplete where there was no
'if self.q' clause. I'm not sure what issues this caused, but now it's
fixed.