Split SourceTraitSearchForm into super and subclasses
Create a general SourceTraitSearchForm which has fields for trait
name and description. Create a second form that is a subclass called
SourceTraitSearchMultipleStudiesForm, which adds the studies field.
Use the second form in the existing SourceTraitSearchView, with the
plan to use the second form when searching for traits within a
a single study.