Add a page to search phenotypes from a specific study
Add a new view to display the form to search traits in a specific
study using the updated (superclass) SourceTraitSearchForm. Add a
new url pattern for this page. Update the Study model's function
to get the search page url. Add a new html template for this page.
Add tests for the new view and update the Study model test for
get_search_url.