Add a SearchFormMixin for use in search views
Add a mixin the logic in the get, form_valid, and form_invalid
methods, plus a search method that subclasses must override. This
removes code that was almost entirely duplicated between the
different search views. Use the new mixin in the search views, and
change them to subclass TemplateView instead of FormView because
SearchFormMixin subclasses FormMixin.