Update forms for searching in specific study datasets
Add a new form subclass of SourceTraitSearchForm specifically for
searching in a study. Require the study as an argument to the form
so that the dataset autocomplete view works properly (since it
requires the study pk to generate the url). Add tests for the new
form. Ensure that the selected dataset both belongs to the study in
question and that it is not deprecated. Add tests for the form
behavior.