Improve tests of user-specific view behavior
Tests of user-specific view elements (e.g. the tagging buttons showing
up only for some users) used to use assertContains and assertNotContains
with strings for e.g. the button names. But that broke as soon as I made
changes for terminology consistency. Rewrite these tests to instead look
for links to the appropriate pages in the assertContains and
assertNotContains tests.