Fix tooltip js file
Tooltips on the source trait detail page (for the little x button next
to a tag, for tag removal) were not working. I didn't realize I had to
add the tooltip.js file and that it wasn't included with bootstrap. So
add the tooltip.js file. Also make the tooltip prefer to show up on the
right.