Fixes after review
Just a couple of tiny fixes after reviewing this branch in preparation
for merging. Add "Test" to the end of a test class name. Check the
staging server data for datasets with names completely composed of
digits and update the note about this check in trait_browser.views'
StudySourceDatasetNameOrPHTAutocomplete class.