Remove short words when searching variable descriptions
Remove words less than 3 characters in the trait search forms
because those words are not indexed, and existing behavior could
cause confusion if users expect them to be. Add a custom field with
a custom clean method to use for the description field in the trait
search forms. Use this field in the forms for fields that will be
passed to the watson search functions. Update the clean methods to
account for the fact that these fields might not be in the data
dictionary, which happens if they do not pass their custom clean
methods. Add tests to check that short words are being properly
removed.