Improve h. trait detail page, esp. uks
Redo the layout of the harmonized trait set version detail view. Get the
linked traits in the view's context, rather than in the template. Get
the main trait and the unique key traits separately. Then display the
main trait just like the source traits are and have a collapsed section
with info panels for each of the u. k. traits and their encoded values.
Remove the Documentation: Coming soon nonsense.