Fix build_test_db errors
Noticed a few errors in the build_test_db function while I was using it
to write some one-off code to create a harmonized trait set that has
unique keys. Set one of the harmonized traits to have i_is_unique_key
True. Remove an extraneous for loop in creation of encoded values.