Add a message with omitted words in search terms
Display a message on the search page listing any entered terms that
were omitted from the watson search for being too short. Add a field
to the WatsonSearchField class that tracks a warning message, and
display it on the search pages if it has been populated with a valid
message. Add tests to forms and search views for this behavior.