Add dataset fields to form for searching source traits
Add fields for the dataset name, description, and a boolean checkbox
of whether to match on exact name to the form used for searching
source traits. Update the layout to display the fields in a similar
way to the trait property fields. Add tests to see if the form
is valid. Update the view to process these fields properly and to
filter source traits by specified dataset properties.