Use shared layout objects in search forms
Add one layout to display the name, match_exact_name, and description
fields in a form. This has two benefits: 1) if we want to change the
layout for these fields in the future, we only have to change it in
one place, and 2) we can wrap all fields in the same col-sm-10 and
col-sm-off-set-1 divs, instead of wrapping fields separately. Add a
second layout to display the search and reset buttons, since these
should be the same across all search forms/views.