Update form layout so messages are properly displayed
Add a Row object to the form layouts to make the grid system work
properly. With the new form layout, the search result message was
appearing on top of the form. Adding a div with class="row" seemed
to fix it (and also implied in bootstrap3 docs) but I am not sure
why it started happening now when it didn't before. It started in
commit d91e23c25cff2e10fe4a7a2536f7ad3466348ecf.