Force select2-containers to be 100% width
The django autocomplete select2 containers were not reliably
rendering at the proper width. Add custom css to set the width of
css class='select2-container' to be 100%. If we need the width of
a container to be less, we can just wrap it in a div of the desired
size.

root? github issue: https://github.com/select2/select2/issues/3278