Add a collapsible panel on trait search page
Hide some of the more advanced filters on the trait search page by
putting them in a collapsible panel. Add logic to the
SourceTraitSearchMultipleStudiesForm so determine if the panel
should be active by default (if data was passed to one of the
contained fields) or inactive by default (if data was not passed to
one of the contained fields).