Fix LoginRequiredTestCase for pks
I'd been waiting on the Django 1.9 upgrade to use resolve(url).func.view_class()
in order to get from the url to the view class to the model class, to improve
this test (which fails in its current form with MariaDB in some tests). But
this didn't turn out to be a good way to implement the test. It didn't account
for any function-based views at all, and didn't account for views that don't have
the model attribute. Instead, I just added a while loop that increments the pk
until it finds one that exists.

See corresponding commit d243bd1 in phenotype_visit_recorder.