Fix object lookup views
The views for object lookup by name or phs didn't work after the upgrade
to Django 1.9. This includes StudyLookup, SourceDatasetLookup, and
SourceTraitLookup. Upon further inspection, these views were all missing
calls to Super().get_context_data in their get_context_data methods. The
fact that they'd worked before was kind of a fluke, but with the Django
1.9 upgrade the FormView class was modified such that the form is added
to the context in get_context_data (not sure where it was happening
before). Add calls to Super().get_context_data in each of the
get_context_data methods to fix this.