Add on_delete settings
In Django 2.0, the on_delete argument is required for all ForeignKey and
OneToOneField fields. Add on_delete arguments to all of these fields in
both models.py and migrations files. For ForeignKey fields that are
imported from topmed_pheno, Adrienne and I decided to use CASCADE in
order to make it easier if we need to delete anything from PIE later on.
We might have used PROTECT for the component trait ManyToMany fields, as
this is comparable to the setting used in topmed_pheno, except that
Django's ManyToManyField uses CASCADE without a way to change the
setting, except by adding a through model.