Fix links to component harmonized traits
HarmonizedTrait's get_name_link_html() method was accidentally using
the pk for the trait rather than the pk for the trait set version
to link to the appropriate component trait set version. Change the
code to use the trait set version pk. This requires fill_fields to be
run again.