Remove a migration dependency
In commit 64f66a I added a dependency on ('sites', '__latest__') to the
core app's migration 0002, which sets up initial groups. This was based
on an error I was getting during the Django 1.9 upgrade process. But now
I can't reproduce the error in Django 1.11 and instead I'm seeing an
InconsistentMigrationHistory error when I try to migrate, because my
migration files say that core/migrations/0002_add_default_groups is
dependent on the sites 0002_alter_domain_unique migration, but my
migration history in the staging database ran them in an order that is
inconsistent with that dependency. Removing this dependency doesn't break
any tests or produce the error I was seeing during the upgrade process,
so I'm going to remove it.