Update assert_redirect_all_urls to work with multiple pks
Update assert_redirect_all_urls to work properly with an arbitrary
number of pks in the url, assuming they are named like "<pk...>". If
they are named "<something else>", the test will fail. Note: before
this change, the test could result in an infinite loop if there
were multiple regex patterns in the url if one of them was <pk>.