Add initial view definition for tag-study table
Add the definition of the view for the list of tagged traits by tag
and study, although it is not finalized yet. Add a test for regular
non-tagger and non-admin users. Update the template to display the
page, with the idea that aesthetic updates will be made later. Add
a subtitle block to the __list.html template so it can be included
on this page.