Add a delete button to TaggedTraitDetail for some groups
Display a delete button on TaggedTraitDetail page for dcc staff
and phenotype taggers from that trait's study. Add new test cases
for dcc staff users and phenotype tagger users. Add a mixin for
TaggedTraitDetail tests and use it in the three different TestCases
(regular users, phenotype taggers, and dcc staff) so that the same
tests are run for all three TestCases without having to duplicate
the code.