Add form errors for existing archived TaggedTraits
For all of the tagging forms, add errors to the appropriate field when
trying to create an existing (but archived) tagged trait. For now, these
errors are the same as for an existing (but not archived) tagged trait,
but they will be changed later, once study users don't see archived
traits.