Add mixin to be used with DCCReview creation or updates
Many of the class-based view methods for reviewing/updating a
TaggedTrait by itself will be the same as those for adding a new
DCCReview in the planned loop setting. To facilitate adding the
ability to review a single TaggedTrait in the future, add a mixin
with much of the shared code. (The specifics of what's in this
mixin may have to change in the future.)