Add a first-pass series of views for reviewing tagged traits
Add three views for reviewing tagged traits, which provide an
interface for the DCC analysts to review multiple traits one after
another, in a kind of loop. Use session variables to store the
necessary info for this loop to work.

The views are:
1. A view to select which tag and study to review, using the
   TaggedTraitReviewSelectForm. This view sets a number of session
   variables with values taken from the form: the study pk, the
   tag pk, and the set of tagged traits to review.
2. A view to redirect users to the appropriate next page, depending
   on whether the session variables have been set and if so, their
   values. If the appropriate session variables have not been set,
   the user is redirected to the form to select a study/tag combo.
   If the session variables have been set, the user is redirected to
   the next TaggedTrait to review (setting another session variable
   that will be used by the next view) or to the list of
   TaggedTraits for that study and tag (if no TaggedTraits are left
   to review).
3. A view to review a tagged trait, and redirect to view 2 above.

The basic structure and setup of the views is working, but tests
have not been written. The views will most likely be modified as
tests are added in the next series of commits.