Redirect to next view if tagged trait has been reviewed
Instead of failing with a database IntegrityError when attempting to
review a TaggedTrait that already has an associated DCCReview,
redirect to the next view and display and informative message
(except if the analyst clicked on the "skip" button). Also add a
helper method to the view that updates the session variables
appropriately to avoid duplicated code, since they need to be
updated in a couple of methods. Add tests.