Redirect if session variables are not appropriately set
In both the "next" and "review" views, check if the expected session
variable is set, and redirect to the "select" view if it is not.
Additionally, in the "review" view, check if 'pk' is a key in the
session variable and redirect to the "next" view if it is not. Add
tests.