Filter study choices in form by the selected tag
Only display studies that have a trait tagged with the selected
tag in the review select form. To accomplish this, modify the
autocomplete view to accept a "forwarded" parameter and then
forward the appropriate parameter in the study form field's
django-autocomplete-light widget. Add tests to check that this is
working appropriately.