Add custom js to clear the study field if new tag is selected
If the user selects a tag, selects a study, and then selects a new
tag, there is no guarantee that the previously-selected study would
still have been a choice (due to forwarding the tag to the study
autocomplete method). Add a javascript file that will hold custom
js for this page. Include this file in the form's media. Update the
form's help text to clarify the order that the fields should be
selected.

Code for javascript to clear the field upon changing the first can
be found here:
https://django-autocomplete-light.readthedocs.io/en/master/tutorial.html#building-blocks-for-custom-logic