Raise form error if no tagged traits exist for study/tag combo
Add a custom clean method to the review select form if the chosen
study/tag combo does not have any unreviewed tagged traits. Note
that this should never happen, due to the study dropdown filtering
added in the previous commit, but it is a good backup. Update the
views and test to remove the warning message that was previously
shown when this occurred.