Add a mixin that processes session variables
Add a mixin to process session variables. The processing can be
checking and (if necessary) redirecting or raising a 404 error, or
setting class attributes to specific values. Adding this mixin
makes the redirection due to the LoginRequiredMixin happen before
the redirection due to improperly set session variables, as long as
the LoginRequiredMixin comes before the SessionVariableMixin.