Prevent users from deleteing confirmed TaggedTraits
Add a new mixin that validates some property of an object and
redirects the user to a different page upon failure. Use this mixin
in the TaggedTraitDeleteView and check that it has not already been
confirmed by the DCC. It would have been possible to override the
delete method of DeleteView to have the same behavior, but in that
case the user was taken to the confirmation page and shown the error
message only after confirming. This way, they are redirected before
even seeing the confirmation page. I had to use a new mixin in order
to allow the login checks from LoginRequiredMixin to run before the
delete validation so that the view passed the LoginRequiredTest
(same reason as behind the SessionVariableMixin).