Rework if statement when deleting tagged trait
The if hasattr() statement was missing an else clause, because
two checks must pass to force the tagged trait to not be deleted:
first, it must have been reviewed and second, the review status
must be confirmed. Instead of having two separate if statements,
combine them into one if statement.