Rename "delete" field in tag-related django-tables2 tables
Rename the "delete" field to "delete_button" in django-tables2
tables that display a button to delete a tagged trait. A field named
delete will cause issues when we override the delete method,
because that method will be called every time the table is
rendered (thus deleting all tagged traits in the table!).