Do not allow any reviewed TaggedTrait to be deleted
Override the delete method in order to prevent a TaggedTrait that
has been reviewed to be deleted, regardless of its status. Add a
custom exception that is raised in this case; it can be used in
the future if there are other objects that we (conditionally) don't
want to be deleted.