Do not allow reviewed TaggedTraits to be deleted in delete view
Check that the TaggedTrait trying to be deleted only has a dcc_review
instead of also looking at its status, and if so do not allow the
delete to happen. Rename the constant storing the error message so
it is clearer (REVIEWED instead of CONFIRMED). Update the tests for
TaggedTraitDelete view so check that reviewed traits that need
followup cannot be deleted.