Improve StudyTaggedTraitList view
In the __study_page.html template, only display the button linking to
the StudyTaggedTraitList view if there there are existing taggedtraits
for the study. In the study_taggedtrait_list.html template, include text
saying that there are no tagged variables for the study when the number
of taggedtraits is 0. Move the study_taggedtrait_list.html template from
the templates/tags directory to templates/trait_browser. Update the
tests of StudyTaggedTraitList for the recent changes.