Add "Continue reviewing" link to the navbar
Add a link to the navbar that is only shown if the user is in the
middle of reviewing a set of traits that allows them to continue
reviewing them without going through the tag/study selection again.
Add tests to check that this link only shows up for users with
permission to add new DCC reviews and only when the appropriate
session variable is set.