Prepare DCCReview forms for new subclassing
Rename DCCReviewCleanForm to DCCReviewBaseForm. Move the form Meta
definition from DCCReviewForm to DCCReviewBaseForm and specifically
use the superclass's Meta in DCCReviewForm. Move the definition of
the SUBMIT_CONFIRM and SUBMIT_FOLLOWUP attributes into the base
form class, which requires defining the layout for the subclass in
an init method. Move tests into a mixin and use it to test
DCCReviewForm. The DCCReviewBaseForm will be subclassed to create a
form for reviewing a single tagged trait. The only difference
between the two is the set of buttons that will appear at the
bottom of the form.