Move info message in review loop out of template
Remove the info box at the top of the template for the DCCReview
form, and add it as a message in the TaggedTraitreveiwByTagAndStudyNext
view. This way, the template can be used for both the reviewing
loop and creating/updating a single DCCReview, and the message about
the number of tagged traits left to review will only appear when
appropriate.