Turn off active study tagged traits tab on profile
Make the "Tagged variables in my studies" tab inactive. This makes all
of the tabs inactive for every user at first view. The alternative is to
do some elaborate if testing in the view to determine which tab should
be active, when there's a large number of possible types of users. So
just making them all inactive is easiest.