Redirect intelligently after TaggedTraitDelete
Modify TaggedTraitDelete view to send the user back to the page they
came from, when it makes sense. Add code to the get_context_data and
get_success_url methods to make this happen. Also add '?next=' to each
link to the TaggedTraitDelete view from other views (in the templates).
Make the cancel button on the confirm_delete template send the user back
to the previous page, if available.