Show the detail link button in all tagged trait tables
Move the definition of the detail column to TaggedTraitTable so that
it is included in all tables and so that it doesn't need to be
redefined in each table. Previously it was only in the tables seen
by the phenotype taggers and DCC staff, but now it is visible for
regular users too.