Display StudyResponse info on TaggedTraitDetail page
Rearrange the get_context_data definition and internal variables to
clarify when the variable refers to a DCCReview. Add a flag that
specifies whether to display the STudyResponse info on the page.
Update the tests. Update the template to show the StudyResponse
info, similar to the DCCReview info. Only show the StudyResponse
info to DCC staff and phenotype taggers for that study. Do not show
buttons to update a DCCReview if a StudyResponse exists.