Add a view to create a disagree StudyResponse
Add a view definition that creates a StudyResponse with disagree
status. This view is similar to the StudyResponseCreateAgree view,
which sets the status to a constant based on the url, but unlike
that view, it requires user input and therefore is a FormView. Add
a Mixin that handles checking whether a TaggedTrait needs a
StudyResponse and redirects appropriately if not. Use this mixin in
both the StudyResponseCreateAgree and StudyResponseCreateDisagree
views. Add a url pattern and a basic template for the new view.