Show agree/disagree buttons in DCCReview table
Add a new table class that includes button to agree or disagree
with a DCCReview. Display this table to phenotype taggers in the
view that shows the table of traits that need study followup.
(Non-taggers and DCC staff do not see the buttons.) Render the
buttons in a template as a form, similar to how the delete view
template works: the "agree" button submits an empty post request to
the StudyResponseCreateAgree view, and the "disagree" button is
simply a link to the StudyResponseCreateDisagree view, which
provides a form for the tagger to input a comment.