Include tag panel in needs followup list page
Include the tag panel in the template rendering the DCCReviewTable
for TaggedTraits that need study followup. Remove the tag column
from the table since the all traits have the same tag, and the tag
information is included in the panel. To allow this, add two new
blocks to the __list.html template: before_table and after_table.