Add view and form to create a StudyResponse
Add a form for creating a study response. Add a new view and
template that processes this form, with a new url pattern. Add a
mixin to help with this that will also be useful for the view to
update an existing StudyResponse. Move the failure url in the
StudyResponseCheckCreateMixin to a method that must be defined
by any class using it.