Rework CBVs for StudyResponseCreate/Update views
Remove the StudyResponseCheckCreateMixin and handle checking in
the individual class based views that create StudyResponse objects,
because they will have different behavior if a StudyResponse exists
already. StudyResponseCreate redirects to StudyResponseUpdate, while
StudyResponseCreateAgree and StudyResponseCreateDisagree redirect
back to the list of TaggedTraits that need study responses. It got
too complicated to try to handle the different behavior with mixins.