Register StudyResponse in admin
Register the StudyResponse model in the admin interface. Add a form
that can be used to do custom cleaning in the admin using the same
base form subclass as the form on the main site. Add a form test
mixin that contains the tests that used to be in
StudyResponseFormTest, and use this mixin in the test classes for
the main form and for the admin form. Also add a str method for
DCCReview so that it renders more nicely in the admin.