Update TaggedTrait detail page wording and aesthetics
Add a dropdown panel for showing quality review status to study
taggers and dcc staff. Include buttons that allow a study tagger to
add a response when appropriate. Make the DCC buttons to add or
update a DCCReview smaller. Move these buttons into the quality
review panel. Add and update existing context flags to match what
needs to be shown in the TaggedTraitDetail view.