Show response status in column of StudyResponse buttons
In the list of TaggedTraits that need followup from a given tag and
study, show buttons only for those TaggedTraits that still need a
StudyResponse. If a StudyResponse already exists, show a disabled
button indicating the status. Update the template with explanatory
information. This change was necessary because the table now shows
all TaggedTraits that need followup whether or not they have a
StudyResponse.