Change url for quality review of tag/study combo
Remove the "reviewed" pattern from the url that phenotype taggers
will go to to respond to TaggedTraits that have been flagged for
removal. This is because this page is specific to study users; we
may want separate pages to show tagged traits with different types
of review status for dcc or study users in the future, and the
"reviewed" pattern would make more sense there.