Handle archived tagged traits in DCCReviewByTagAndStudy
In the DCCReviewByTagAndStudy view, add an if statement to check whether
the tagged trait about to be reviewed has already been archived in the
meantime. If it has, give a warning message and redirect to the
DCCReviewByTagAndStudyNext view to get the next pk for reviewing. Fix
tests for this view and archived tagged traits as well.