Fix study response buttons in table
Fix rendering of the study response buttons in the
DCCReviewTableWithStudyResponseButtons table. The "remove" button
broke in commit bc80e2c2ae61e5a30cc31b49b1a62ef341b7b96a because
the csrf_token was not properly rendered with get_template.render().
Instead of overwriting the render_ method for the buttons column,
add the logic to the template and use a TemplateColumn. Add tests
that check for the presence of a csrf token and that the buttons
are being displayed when appropriate to the views tests. Remove
the tests of whether the buttons are displayed from the tables
tests, beacuse the template needs a request object to render
properly (which is not present in any of the tables tests).