Handle archived tagged traits in DCCReviewUpdate
Majorly revise DCCReviewUpdate to accommodate archived tagged traits.
Move the try/except that checks for the existences of a DCCReview object
into get_object(). Rewrite get() and post() because the checking for
archived tagged traits needs to happen in the middle of what happens in
UpdateView.get() and .post(). These methods basically run get_object()
and then super().get() or super().post(), but the checks need to happen
after get_object() but before the super() call. So rewrite them to do
this. Use the ProcessFormView.get() and ProcessFormView.post() methods
because that is what UpdateView ultimately inherits from. Update tests
for this new behavior.