Add count of reviewed tagged traits to quality review table
Add a second column to the tag/study quality review table that
displays the number of tagged traits that have already been reviewed
for that tag. Change table headers to be more clear about what the
two numbers in the table are. Only display the "begin quality
review" button when there is at least one tagged trait that needs
a study response, and display an informative message otherwise. Add
tests to check the behavior.

Note: the django docs talk about a bug when you try to add two new
columns with annotate (which is what is being done here), but I
think it is only a problem if the joins produced by the sql query
result in duplicated rows in a table. Because none of the joins
produce duplicated tagged traits, it should not be a problem.